TUC-128 Added a CycleStatisticsCalculator.
More elegant than the previous solution.

Also a lot of small improvements, thanks to the
test of ConstantElimination.

	modified:   make/qtcreator/tucucore.pri
	modified:   src/tucucore/computingcomponent.cpp
	modified:   src/tucucore/computingservice/computingresponse.h
	modified:   src/tucucore/computingservice/computingtrait.cpp
	modified:   src/tucucore/computingservice/computingtrait.h
	new file:   src/tucucore/cyclestatistics.cpp
	new file:   src/tucucore/cyclestatistics.h
	modified:   src/tucucore/cyclestatisticscalculator.cpp
	modified:   src/tucucore/cyclestatisticscalculator.h
	modified:   src/tucucore/drugmodel/analyte.h
	modified:   src/tucucore/generalextractor.cpp
	modified:   src/tucucore/generalextractor.h
	modified:   src/tucucore/invariants.h
	modified:   src/tucucore/montecarlopercentilecalculator.cpp
	modified:   src/tucucore/residualerrormodel.cpp
	modified:   src/tucucore/residualerrormodel.h
	modified:   src/tucucore/unit.h
	modified:   test/tucucore/drugmodels/buildconstantelimination.h
	modified:   test/tucucore/drugmodels/test_constanteliminationbolus.h
	modified:   test/tucucore/test_cyclestatistics.h

+++ test/tucucore/test_cyclestatistics.h
@@ -30,8 +30,8 @@ struct TestCycleStatistics : public fructose::test_base<TestCycleStatistics>
     {
         // set cycleData
         DateTime now;
-        int nbCompartments = 1;
-        int nbPoints = 10;
+        size_t nbCompartments = 1;
+        size_t nbPoints = 10;
         Tucuxi::Core::CycleData cycleData;
         std::vector< std::vector<Tucuxi::Core::CycleStatistic> > stats;
 
@@ -40,7 +40,7 @@ struct TestCycleStatistics : public fructose::test_base<TestCycleStatistics>
         std::vector<Concentrations> concentrations{{0, 10.2, 20.5, 15.0, 9.7, 10, 12.7, 15.2, 13.2, 14}};
         std::vector<TimeOffsets> times{{0, 0.5, 1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5}};
 
-        for (int compartment = 0; compartment < nbCompartments; compartment++) {
+        for (size_t compartment = 0; compartment < nbCompartments; compartment++) {
             // assign concentration data
             cycleData.m_concentrations.push_back(Concentrations ());
             cycleData.m_concentrations[compartment].reserve(nbPoints);
